+++ CHANGELOG.md
@@ -82,3 +82,4 @@ October 30th, 2022
     - Data points, file names, and save directories can be customized in the configuration.
 - Added a configuration option to disable diagonal arrows.
     - When diagonal arrows are disabled, Assassin will point in increments of 90 degrees instead of 45.
+- Removed relay alert logic.